Fix iso's larger than 4GiB

+++ iop/common/cdvd_config.h
@@ -39,7 +39,7 @@ struct cdvdman_fragfile
 {
     u8 frag_start; /// First fragment in the fragment table
     u8 frag_count; /// Number of fragments in the fragment table
-    u32 size;      /// Size of the file in bytes
+    u64 size;      /// Size of the file in bytes
 } __attribute__((packed));
 
 struct cdvdman_settings_bdm
